+++ include/mockturtle/algorithms/node_resynthesis/mig_npn.hpp
@@ -52,10 +52,11 @@ namespace mockturtle
 
 /*! \brief Resynthesis function based on pre-computed size-optimum MIGs.
  *
- * This resynthesis function can be passed to ``node_resynthesis``.  It will
- * produce an MIG based on pre-computed size-optimum MIGs with up to at most 4
- * variables.  Consequently, the nodes' fan-in sizes in the input network
- * must not exceed 4.
+ * This resynthesis function can be passed to ``node_resynthesis``,
+ * ``cut_rewriting``, and ``refactoring``.  It will produce an MIG based on
+ * pre-computed size-optimum MIGs with up to at most 4 variables.
+ * Consequently, the nodes' fan-in sizes in the input network must not exceed
+ * 4.
  *
    \verbatim embed:rst